RED: Flesh out Response type
GraphQL spec provides a limited range of things it can be. This is now
encoded in `Response`. Breaks handlers because they currently expect
return type of "resolvers" (not necessarily the same as the technical
term "resolver" from GraphQL spec, merely our stop-gap definition) to be
identical to the return type of the whole request.