Flesh out Response type
GraphQL spec provides a limited range of things it can be. This is now
encoded in `Response`.

Doing this makes it obvious that the return value of the top-level
process cannot be the same as that of the individual "handlers". It also
makes it obvious that we're allowing non-leaf nodes to be queried, which
is against the spec.