Implement most of query validation
The `Validation` module now contains a structured `QueryDocument` data type
ready for use by the executor.

Scope of the module is defined by the
[Validation](https://facebook.github.io/graphql/#sec-Validation) chapter of
the spec, except as otherwise noted.

Over-arching philosophy is to do "constructive" validation, making new objects
that are the ones we wish we had, and reporting errors where not found. Thus,
we use Maps to show that names are unique.

There is still more to do, but this is big enough and complete enough that it
should be reviewed & merged.