Allow Handlers to return errors.
Prior to this change, the graphql-api internals could generate errors
but the user-supplied Handler could not.  This adds the ability for
the user-supplied Handler to also return an error.

1. Added a HandlerError constructor to the ResolverError to represent
   errors generated by the handler.

2. Updated the resolve to return a 'HandlerResult a' which is now a
   typedef for 'Either Text a'.  This could not be done as a 'Union'
   type because the 'Left err' value should have a different effect on
   the results (null in 'data' and adding to 'errors').

3. Updated all the HasResolver class instances for the change in #2.

4. Added 'returns' and 'handlerError' convenience functions for
   handlers to abstract the internals of the success/failure
   implementation.

5. Enhanced the tests (particularly ResolverSpec) for testing the
   generation of success and error results for different handler
   types.

This change is *not* backward compatible---and cannot be because the
handlers now need the ability to signal errors.  However, all existing
handlers which perforce return successful values can be updated rather
simply using the new 'returns' helper for each handler field instead
of the previous 'pure' function:

    miniCat name = pure (pure name :<> pure 32)

becomes:

    miniCat name = pure (returns name :<> returns 32)

And:

    handler :: Handler IO Query
    handler = pure $ \fooId -> do
       foo <- lookupFoo fooId
       -- note that fmap maps over the Maybe, so we still need
       -- have to wrap the result in a pure.
       sequence $ fmap (pure . viewFoo) foo

becomes:

    handler :: Handler IO Query
    handler = pure $ \fooId -> do
       foo <- lookupFoo fooId
       returns $ viewFoo <$> foo

The current reported package version of 0.3.0 should be changed to at
least 0.4.0 to indicate a breaking, non-backward-compatible change.